+++ src/boot/stage_2/memory/memory.h
@@ -4,4 +4,4 @@
 
 bool memcmp(const void *lhs, const void *rhs, unsigned int count);
 
-void * memcpy ( void * destination, const void * source, uint64_t num );
+void * memcpy ( void * destination, const void * source, uint32_t num );
